refactor(notion): migrate to bundled plugin manifest
