+++ backend/app/plugins/notion/__init__.py
@@ -1,21 +1,1 @@
-"""Notion integration for Pawrrtal.
-
-Importing this package registers the Notion :class:`Plugin` against
-:mod:`app.agents.plugins.registry`, exposing a single ``notion_cli``
-tool to any agent whose workspace has a ``NOTION_API_KEY`` configured.
-
-Execution is delegated to the official Notion CLI (``ntn``):
-:mod:`app.plugins.notion.ntn_client` shells out per call with the
-workspace's token injected via ``NOTION_API_TOKEN`` and an isolated
-``HOME`` so any state ``ntn`` writes can't leak between workspaces.
-Each invocation is logged to :class:`app.models.NotionOperationLog`
-so operators can audit what the agent ran without scraping uvicorn
-output.
-
-Design rationale lives in
-``frontend/content/docs/handbook/decisions/2026-05-15-plugin-system-and-notion-integration.mdx``.
-"""
-
-from app.plugins.notion.plugin import notion_plugin
-
-__all__ = ["notion_plugin"]
+"""Notion runtime helpers used by the manifest-backed Notion plugin."""
